refactor: update docs & minor changes in runtime
