+++ .changeset/orange-monkeys-obey.md
@@ -1,0 +1,5 @@
+---
+'@scale-codec/core': patch
+---
+
+Update typings for Enum encode/decode
