+++ packages/docs/root/snippets/namespace-schema-compiled.ts
