+++ packages/docs/root/snippets/namespace-schema.ts
