+++ scripts/compile-docs-namespace.ts
@@ -1,9 +1,9 @@
 import { path } from 'zx';
 import fs from 'fs/promises';
-import definition from '../packages/docs/root/guide/snippets/namespace-schema';
+import definition from '../packages/docs/root/snippets/namespace-schema';
 import { renderNamespaceDefinition } from '../packages/docs/node_modules/@scale-codec/definition-compiler/src/lib';
 
-const OUTPUT_FILE = path.resolve(__dirname, '../packages/docs/root/guide/snippets/namespace-schema-compiled.ts');
+const OUTPUT_FILE = path.resolve(__dirname, '../packages/docs/root/snippets/namespace-schema-compiled.ts');
 
 export default async function () {
     const content = renderNamespaceDefinition(definition);
